langmodel/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
crispy-bootstrap5==2024.10
cryptography==39.0.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.2
django-allauth==65.0.2
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.14
langchain-core==0.3.29
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langgraph==0.2.61
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.48
langsmith==0.2.10
msgpack==1.1.0
multidict==6.1.0
numpy==2.2.1
oauthlib==3.2.2
openai==1.59.5
orjson==3.10.14
packaging==24.2
propcache==0.2.1
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.21
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.6.0
python3-openid==3.2.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.4.3
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.14
whitenoise==6.7.0
yarl==1.18.3
